added ascii mapping and divided cipher utils into multiple files
