+++ GaloisFieldHillCipher/src/Parsing/ArgumentParser.c
@@ -8,7 +8,6 @@ STATUS_CODE extract_arguments(ParsedArguments* out_args, int argc, char** argv)
     const char* key = NULL;
     uint32_t dimension = 0;
     uint32_t number_of_random_bits_between_bytes = 0;
-    const char* output_format_string = NULL;
     uint32_t verbose = false;
     OPERATION_MODE mode = MODE_UNINITIALIZED;
     const char* mode_string = NULL;
@@ -22,7 +21,6 @@ STATUS_CODE extract_arguments(ParsedArguments* out_args, int argc, char** argv)
         OPT_BOOLEAN(*ARGUMENT_VERBOSE_SHORT, ARGUMENT_VERBOSE_LONG, &verbose, ARGUMENT_VERBOSE_DOCUMENTATION),
         OPT_STRING(*ARGUMENT_MODE_SHORT, ARGUMENT_MODE_LONG, &mode_string, ARGUMENT_MODE_DOCUMENTATION),
         OPT_STRING(*ARGUMENT_KEY_SHORT, ARGUMENT_KEY_LONG, &key, ARGUMENT_KEY_DOCUMENTATION),
-        OPT_STRING(*ARGUMENT_OUTPUT_FORMAT_SHORT, ARGUMENT_OUTPUT_FORMAT_LONG, &output_format_string, ARGUMENT_OUTPUT_FORMAT_DOCUMENTATION),
         OPT_INTEGER(*ARGUMENT_RANDOM_BITS_SHORT, ARGUMENT_RANDOM_BITS_LONG, &number_of_random_bits_between_bytes, ARGUMENT_RANDOM_BITS_DOCUMENTATION),
         OPT_END(),
     };
@@ -31,13 +29,6 @@ STATUS_CODE extract_arguments(ParsedArguments* out_args, int argc, char** argv)
     argparse_init(&argparse, options, NULL, 0);
     argparse_describe(&argparse, USAGE, NULL);
     if (argparse_parse(&argparse, argc, (const char**)argv) < 0 || !mode_string)
-        goto parse_error;
-
-    if (!output_format_string || strcmp(output_format_string, "binary") == 0)
-        output_format = OUTPUT_FORMAT_BINARY;
-    else if (strcmp(output_format_string, "text") == 0)
-        output_format = OUTPUT_FORMAT_TEXT;
-    else
         goto parse_error;
 
     if (strcmp(mode_string, MODE_KEY_KEY_GENERATION_SHORT) == 0)
@@ -61,26 +52,22 @@ STATUS_CODE extract_arguments(ParsedArguments* out_args, int argc, char** argv)
     switch (mode)
     {
     case KEY_GENERATION_MODE:
-        if (!output_file || !dimension ||
-            STATUS_FAILED(validate_output_file(output_file, output_format)))
+        if (!output_file || !dimension)
             goto parse_error;
         break;
 
     case DECRYPTION_KEY_GENERATION_MODE:
         if (!key || !output_file || !dimension ||
             STATUS_FAILED(validate_file_is_readable(key)) ||
-            STATUS_FAILED(validate_file_is_binary(key)) ||
-            STATUS_FAILED(validate_output_file(output_file, output_format)) )
+            STATUS_FAILED(validate_file_is_binary(key)))
             goto parse_error;
         break;
 
     case DECRYPT_MODE:
         if (!input_file || !output_file || !key || !dimension ||
             STATUS_FAILED(validate_file_is_readable(input_file)) ||
-            STATUS_FAILED(validate_file_is_binary(input_file)) ||
             STATUS_FAILED(validate_file_is_readable(key)) ||
-            STATUS_FAILED(validate_file_is_binary(key)) ||
-            STATUS_FAILED(validate_output_file(output_file, output_format)))
+            STATUS_FAILED(validate_file_is_binary(key)))
             goto parse_error;
         break;
 
@@ -88,16 +75,14 @@ STATUS_CODE extract_arguments(ParsedArguments* out_args, int argc, char** argv)
         if (!input_file || !output_file || !key || !dimension ||
             STATUS_FAILED(validate_file_is_readable(key)) ||
             STATUS_FAILED(validate_file_is_binary(key)) ||
-            STATUS_FAILED(validate_file_is_readable(input_file)) ||
-            STATUS_FAILED(validate_output_file(output_file, output_format)))
+            STATUS_FAILED(validate_file_is_readable(input_file)))
             goto parse_error;
         break;
 
     case GENERATE_AND_ENCRYPT_MODE:
         if (!input_file || !output_file || !key || !dimension ||
             STATUS_FAILED(validate_file_is_readable(input_file)) ||
-            STATUS_FAILED(validate_file_is_binary(key)) ||
-            STATUS_FAILED(validate_output_file(output_file, output_format)))
+            STATUS_FAILED(validate_file_is_binary(key)))
             goto parse_error;
         break;
 
@@ -105,9 +90,7 @@ STATUS_CODE extract_arguments(ParsedArguments* out_args, int argc, char** argv)
         if (!input_file || !output_file || !key || !dimension ||
             STATUS_FAILED(validate_file_is_readable(key)) ||
             STATUS_FAILED(validate_file_is_binary(key)) ||
-            STATUS_FAILED(validate_file_is_readable(input_file)) ||
-            STATUS_FAILED(validate_file_is_binary(input_file)) ||
-            STATUS_FAILED(validate_output_file(output_file, output_format)))
+            STATUS_FAILED(validate_file_is_readable(input_file)))
             goto parse_error;
         break;
 
@@ -121,12 +104,11 @@ STATUS_CODE extract_arguments(ParsedArguments* out_args, int argc, char** argv)
     out_args->verbose = verbose;
     out_args->mode = mode;
     out_args->key = key;
-    out_args->output_format = output_format;
     out_args->number_of_random_bits_between_bytes = number_of_random_bits_between_bytes;
+    out_args->output_format = STATUS_SUCCESS(validate_file_is_binary(output_file)) ? OUTPUT_FORMAT_BINARY : OUTPUT_FORMAT_TEXT;
     out_args->input_format = STATUS_SUCCESS(validate_file_is_binary(input_file)) ? OUTPUT_FORMAT_BINARY : OUTPUT_FORMAT_TEXT;
 
     return STATUS_CODE_SUCCESS;
-
 parse_error:
     fprintf(stderr, USAGE);
     return STATUS_CODE_PARSE_ARGUMENTS_FAILED;
